python/ldap_server/requirements.txt
python-msgraph
azure-identity
ldap3